requirements.txt
streamlit
google-generativeai
duckduckgo-search
gTTS
Pillow
streamlit-mic-recorder
SpeechRecognition
PyPDF2